Update master to calculate weaken time correctly
Weaken time needs to use the actual sec level instead of the minimum as it is the action that gets the target to the minimum. Also added the quick to git as a script to quickly test small things instead of test which is more often used for bigger poc code.